update command file names to camelCase
