+++ commands/roles/addRole.js
